# public/swagger.yaml
openapi: 3.0.3
info:
  title: CookMate API
  description: |
    CookMate API - A comprehensive cooking and recipe management platform.
    
    ## Features
    - User authentication with Google OAuth and Email OTP
    - Recipe search and management
    - Ingredient tracking and pantry management
    - Shopping list functionality
    - Favorites system
    - User profile management
    - Meal planning
    - Notifications system
    - Account recovery
    
    ## Authentication
    Most endpoints require Bearer token authentication. Get your token by calling the authentication endpoints.
    
    **Authentication Required Endpoints:**
    - `/api` - API Welcome message
    - `/api/ai` - AI Chat Assistant (OpenAI GPT-4o-mini with tool calling)
    - `/api/admin/users` - Admin: User management (requires admin role)
    - `/api/ingredients` - All CRUD operations
    - `/api/ingredient-categories` - All CRUD operations
    - `/api/favorites` - All CRUD operations  
    - `/api/profile` - All CRUD operations
    - `/api/profile/delete` - Delete account
    - `/api/shopping` - All CRUD operations
    - `/api/recipes/search` - Recipe search
    - `/api/recipes/today` - Today's recipes
    - `/api/meal-plans` - All CRUD operations
    - `/api/notifications` - All CRUD operations
    
    **No Authentication Required:**
    - `/api/auth/google` - Google authentication
    - `/api/auth/otp` - Send OTP for email authentication
...
    - `/api/auth/recover` - Send recovery OTP
...
    
...
    
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/ai:
    post:
      tags:
        - AI Assistant
      summary: Chat with CookMate AI Assistant
      description: |
        Send messages to CookMate AI Assistant for help with recipes, meal planning, ingredients, shopping lists, and more.
        The AI can call various tools to help you manage your cooking activities.
        
        **Available AI Tools:**
        - Profile: get_profile
        - Favorites: get_all_favorites
        - Meal Plans: get_all_meal_plans
        - Ingredients: get_all_ingredients
        - Ingredient Categories: get_all_ingredient_categories, add_ingredient_category
        - Shopping List: get_shopping_list, add_shopping_item
        
        **Example Queries:**
        - "Show my shopping list"
        - "Add milk to shopping list"
        - "What are my favorite recipes?"
        - "Show my meal plans"
        - "Create category Vegetables with icon 🥬"
        - "Show my profile"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - messages
              properties:
                messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                        enum: [user, assistant, tool]
                        description: Message role
                      content:
                        type: string
                        description: Message content
                      toolInvocations:
                        type: object
                        description: Tool invocations (for assistant messages with tool calls)
                  description: Conversation history
                  example:
                    - role: user
                      content: "Show my shopping list"
      responses:
        '200':
          description: AI response generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: object
                    properties:
                      role:
                        type: string
                        example: "assistant"
                      content:
                        type: string
                        example: "Here's your shopping list..."
                      toolInvocations:
                        type: object
                        description: Tool results if AI called tools
                  usage:
                    type: object
                    properties:
                      inputTokens:
                        type: integer
                      outputTokens:
                        type: integer
                      totalTokens:
                        type: integer
                  finishReason:
                    type: string
                    example: "stop"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from authentication endpoints.
        Include the token in the Authorization header as: `Bearer <token>`
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Something went wrong"
  responses:
...
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Something went wrong"
...